refactor: move shared buffer base into ipc
